refactor autograd test
