+++ test/torchaudio_unittest/lfilter_autograd_cpu_test.py
@@ -1,0 +1,9 @@
+import torch
+from .lfilter_autograd_impl import Autograd
+from torchaudio_unittest import common_utils
+
+
+@common_utils.skipIfNoExtension
+class TestAutogradLfilterCPU(Autograd, common_utils.PytorchTestCase):
+    dtype = torch.float64
+    device = torch.device('cpu')
