+++ test/torchaudio_unittest/lfilter_autograd_cuda_test.py
@@ -1,0 +1,10 @@
+import torch
+from .lfilter_autograd_impl import Autograd
+from torchaudio_unittest import common_utils
+
+
+@common_utils.skipIfNoExtension
+@common_utils.skipIfNoCuda
+class TestAutogradLfilterCUDA(Autograd, common_utils.PytorchTestCase):
+    dtype = torch.float64
+    device = torch.device('cuda')
